#4 #7 #8 #3 issue fixes

+++ src/main/java/com/stackroute/muzix/service/TrackServiceImpl.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class TrackServiceImpl implements TrackService {
@@ -26,25 +27,28 @@ public class TrackServiceImpl implements TrackService {
 	
 // 	method for creating a new track DB
 	@Override
-	public boolean saveTrack(Track track) {
-			trackRepository.save(track);
-			return true;
+	public Track saveTrack(Track track) {
+			return trackRepository.save(track);
 	}
 
 // 	method for delete a track from DB using id
 	@Override
-	public boolean deleteTrack(int id) {
+	public Track deleteTrack(int id) {
+		Optional<Track> track =null;
+		if(trackRepository.existsById(id) == true) {
 			trackRepository.deleteById(id);
-			return true;
+			track= trackRepository.findById(id);
+		}
+		return track.get();
 	}
 	
 // 	method for update a already available track in DB
 	@Override
-	public boolean updateTrack(int id, Track track) {
+	public Track updateTrack(int id, Track track) {
 			if(trackRepository.existsById(id) == true){
-				trackRepository.save(track);
-				return true;
+				return trackRepository.save(track);
 			}
-		return false;
+			else
+				return null;
 	}
 }
